Move ISimpleDialog to SimpleDialog
